Moved Text on switch to CatSwitch Added light and dark mode preview fix
